[Test] Split out the test host module definitions.
