+++ test/aot/AOTcoreTest.cpp
@@ -20,6 +20,7 @@
 #include "validator/validator.h"
 #include "vm/vm.h"
 
+#include "../spec/hostfunc.h"
 #include "../spec/spectest.h"
 #include "gtest/gtest.h"
